v0.8.1 - price changes done, also several improvements

+++ app/src/main/java/com/redc4ke/taniechlanie/data/menu/DetailsReviewAdapter.kt
@@ -23,7 +23,7 @@ class ReviewAdapter(
     private val list: List<Review>,
     private val reviewViewModel: ReviewViewModel,
     private val userReview: Review?,
-    private val id: Int):
+    private val id: Long):
     RecyclerView.Adapter<ReviewViewHolder>() {
 
     private val finalList = mutableListOf<Review>()
